ai-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
anyio
